moved nft modal to separate file
