+++ components/QuestChain/QuestChainV1Page/index.tsx
@@ -1,10 +1,6 @@
-import { CopyIcon, ExternalLinkIcon, InfoIcon } from '@chakra-ui/icons';
+import { InfoIcon } from '@chakra-ui/icons';
 import {
   Accordion,
-  AccordionButton,
-  AccordionIcon,
-  AccordionItem,
-  AccordionPanel,
   Alert,
   AlertIcon,
   AlertTitle,
@@ -15,16 +11,13 @@ import {
   Image,
   Input,
   Link as ChakraLink,
-  Link,
   Modal,
   ModalBody,
   ModalCloseButton,
   ModalContent,
-  ModalHeader,
   ModalOverlay,
   Spinner,
   Text,
-  Tooltip,
   useDisclosure,
   useTimeout,
   VStack,
@@ -38,6 +31,11 @@ import { toast } from 'react-hot-toast';
 import { TwitterShareButton } from 'react-share';
 
 import { ConfirmationModal } from '@/components/ConfirmationModal';
+import NFTForm from '@/components/CreateChain/NFTForm';
+import { AwardIcon } from '@/components/icons/AwardIcon';
+import { EditIcon } from '@/components/icons/EditIcon';
+import { PowerIcon } from '@/components/icons/PowerIcon';
+import { TwitterIcon } from '@/components/icons/TwitterIcon';
 import { Page } from '@/components/Layout/Page';
 import { MarkdownEditor } from '@/components/MarkdownEditor';
 import { MarkdownViewer } from '@/components/MarkdownViewer';
@@ -49,6 +47,7 @@ import { QuestTile } from '@/components/QuestTile';
 import { Role } from '@/components/RoleTag';
 import { HeadComponent } from '@/components/Seo';
 import { SubmitButton } from '@/components/SubmitButton';
+import { UploadImageForm } from '@/components/UploadImageForm';
 import { UserDisplay } from '@/components/UserDisplay';
 import { useDropImage } from '@/hooks/useDropFiles';
 import { useInputText } from '@/hooks/useInputText';
@@ -61,15 +60,10 @@ import { waitUntilBlock } from '@/utils/graphHelpers';
 import { handleError, handleTxLoading } from '@/utils/helpers';
 import { Metadata, uploadFiles, uploadMetadata } from '@/utils/metadata';
 import { ipfsUriToHttp } from '@/utils/uriHelpers';
-import { AVAILABLE_NETWORK_INFO, formatAddress, useWallet } from '@/web3';
+import { AVAILABLE_NETWORK_INFO, useWallet } from '@/web3';
 import { getQuestChainContract } from '@/web3/contract';
 
-import NFTForm from '../CreateChain/NFTForm';
-import { AwardIcon } from '../icons/AwardIcon';
-import { EditIcon } from '../icons/EditIcon';
-import { PowerIcon } from '../icons/PowerIcon';
-import { TwitterIcon } from '../icons/TwitterIcon';
-import { UploadImageForm } from '../UploadImageForm';
+import { NFTDetailsModal } from './NFTDetailsModal';
 import { QuestsEditor } from './QuestsEditor';
 import { RolesEditor } from './RolesEditor';
 
@@ -417,11 +411,6 @@ export const QuestChainV1Page: React.FC<QuestChainV1PageProps> = ({
     },
     [refresh, questChain, chainId, provider, onEditNFTClose],
   );
-
-  const copyToClipboard = useCallback((value: string) => {
-    navigator.clipboard.writeText(value);
-    toast.success('Copied to clipboard');
-  }, []);
 
   return (
     <Page>
@@ -1136,98 +1125,11 @@ export const QuestChainV1Page: React.FC<QuestChainV1PageProps> = ({
                     />
                   </Flex>
                 )}
-                <Modal isOpen={isNFTModalOpen} onClose={onNFTModalClose}>
-                  <ModalOverlay />
-                  <ModalContent>
-                    <ModalHeader>{questChain.token?.name}</ModalHeader>
-                    <ModalCloseButton />
-                    <ModalBody>
-                      <Flex mb={3} alignItems="start" direction="column">
-                        <Flex
-                          w="full"
-                          gap={4}
-                          mb={2}
-                          justifyContent="space-between"
-                        >
-                          <NetworkDisplay chainId={questChain.chainId} asTag />
-                        </Flex>
-                        <Image
-                          src={ipfsUriToHttp(questChain.token.imageUrl)}
-                          alt="quest chain NFT badge"
-                          maxW="100%"
-                          maxH="100%"
-                        />
-                        <Accordion allowMultiple w="full">
-                          <AccordionItem>
-                            <AccordionButton>
-                              <Box
-                                as="span"
-                                flex="1"
-                                textAlign="left"
-                                fontSize={13}
-                              >
-                                Additional Info
-                              </Box>
-                              <AccordionIcon />
-                            </AccordionButton>
-                            <AccordionPanel pb={1}>
-                              <Flex
-                                justifyContent="space-between"
-                                w="full"
-                                alignItems="center"
-                              >
-                                <Flex fontSize={12}>
-                                  Address:{' '}
-                                  <Tooltip
-                                    label={questChain.token.tokenAddress}
-                                  >
-                                    <Text fontWeight="bold" ml={2}>
-                                      {formatAddress(
-                                        questChain.token.tokenAddress,
-                                      )}
-                                    </Text>
-                                  </Tooltip>
-                                </Flex>
-                                <Button
-                                  variant="ghost"
-                                  p={0}
-                                  onClick={() =>
-                                    copyToClipboard(
-                                      questChain.token.tokenAddress,
-                                    )
-                                  }
-                                >
-                                  <CopyIcon />
-                                </Button>
-                              </Flex>
-                              <Flex
-                                justifyContent="space-between"
-                                w="full"
-                                alignItems="center"
-                              >
-                                <Flex fontSize={12}>
-                                  Id:
-                                  <Text fontWeight="bold" ml={2}>
-                                    {questChain.token.tokenId}
-                                  </Text>
-                                </Flex>
-                                <Button
-                                  variant="ghost"
-                                  p={0}
-                                  onClick={() =>
-                                    copyToClipboard(questChain.token.tokenId)
-                                  }
-                                >
-                                  <CopyIcon />
-                                </Button>
-                              </Flex>
-                            </AccordionPanel>
-                          </AccordionItem>
-                        </Accordion>
-                      </Flex>
-                    </ModalBody>
-                  </ModalContent>
-                </Modal>
+                <NFTDetailsModal
+                  isOpen={isNFTModalOpen}
+                  onClose={onNFTModalClose}
+                  questChain={questChain}
+                />
                 {isAdmin &&
                   mode === Mode.MEMBER &&
                   !isEditingMembers &&
